Рефакторил выбиралки файлов

+++ cms/reflex/class/controller/storage.php
@@ -17,6 +17,11 @@ class Storage extends Core\Controller {
 	 **/
 	public function post_getWindow($p) {
 	    $editor = \Infuso\Cms\Reflex\Editor::get($p["editor"]);
+        
+        if(!$editor->beforeEdit()) {
+            throw new \Exception("Security error");
+        }
+        
 	    $tmp = app()->tm("/reflex/storage");
 	    $tmp->param("editor",$editor);
 	    return $tmp->getContentForAjax();
@@ -24,6 +29,11 @@ class Storage extends Core\Controller {
 
 	public function post_getFiles($p) {
 	    $editor = \Infuso\Cms\Reflex\Editor::get($p["editor"]);
+        
+        if(!$editor->beforeEdit()) {
+            throw new \Exception("Security error");
+        }
+        
         $storage = $editor->item()->storage();
         $storage->setPath($p["path"]);
         $tmp = app()->tm("/reflex/storage/files-ajax");
@@ -34,19 +44,43 @@ class Storage extends Core\Controller {
 	}
 
 	/**
-	 * @todo проверка безопасности
+	 * Закачивает файл в хранилище для объекта
 	 **/
 	public function post_upload($p) {
+
 	    $editor = \Infuso\Cms\Reflex\Editor::get($p["editor"]);
-	    $storage = $editor->item()->storage();
+        
+        if(!$editor->beforeEdit()) {
+            throw new \Exception("Security error");
+        }
+        
+	    $storage = $editor->item()->storage();        
         $storage->setPath($p["path"]);
-	    $path = $storage->addUploaded($_FILES["file"]["tmp_name"],$_FILES["file"]["name"]);
-	    return $path;
+        
+        $path = null;
+        foreach($_FILES as $file) {
+	       $path = $storage->addUploaded($file["tmp_name"],$file["name"]);
+        }
+        
+	    if($path) {
+            return array(
+                "filename" => $path,
+                "preview150" => (String) \Infuso\Core\File::get($path)->preview(150,150)->fit(),
+            );
+        }
 	}
 
+    /**
+     * Контроллер удаления файла
+     **/         
     public function post_delete($p) {
 
 	    $editor = \Infuso\Cms\Reflex\Editor::get($p["editor"]);
+        
+        if(!$editor->beforeEdit()) {
+            throw new \Exception("Security error");
+        }
+        
 	    $storage = $editor->item()->storage();
         foreach($p["items"] as $file) {
             $storage->delete($file);
@@ -56,6 +90,11 @@ class Storage extends Core\Controller {
 
     public function post_createFolder($p) {
 	    $editor = \Infuso\Cms\Reflex\Editor::get($p["editor"]);
+        
+        if(!$editor->beforeEdit()) {
+            throw new \Exception("Security error");
+        }
+        
 	    $storage = $editor->item()->storage();
         $storage->setPath($p["path"]);
         $storage->mkdir($p["name"]);
